webapp/requirements.txt
Babel==2.5.1
blinker==1.4
click==6.7
coverage==4.4.1
decorator==4.1.2
defusedxml==0.5.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
pbr==3.1.1
python3-openid==3.1.0
pytz==2017.2
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1